Finished first version of Train.java
In this first version, I completed all the requirements. In addition, I added four methods. First, seatsRemainingInEachCar(), which prints the seats remaining situation in each car. Second, welcomeMessage(Passenger p), which welcomes the passenger when they board and show them the remaining seats in each car. Third, selectCar(Scanner scanner), which lets the passenger choose which car to be in. Last, goodbyeMessage(Passenger p), which says goodbye when a passenger gets off the train.